:construction: eliminarGasto removed from ControladorSesion into RepositorioGastos
